requirements.txt
pymupdf>=1.23.0
pdfplumber>=0.10.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
networkx>=3.1
google-generativeai>=0.3.0
pandas>=2.0.0
pyarrow>=12.0.0
orjson>=3.9.0
regex>=2023.0.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
gradio>=4.0.0
python-arango>=7.9.0
